+++ fixtures/parcel-transformer-test-compress-class-name-app/src/index.jsx
@@ -1,0 +1,7 @@
+import '@compiled/react';
+
+const App = () => (
+  <>
+    <div css={{ fontSize: 50, color: 'red' }}>CSS prop</div>
+  </>
+);