+++ packages/babel-plugin/src/styled/__tests__/tagged-template-expression.test.ts
@@ -5,6 +5,10 @@ describe('styled tagged template expression', () => {
   const transform = (code: string, opts: TransformOptions = {}) =>
     transformCode(code, { snippet: true, ...opts });
 
+  const enableTypescript: TransformOptions = {
+    parserBabelPlugins: ['typescript', 'jsx'],
+  };
+
   it('only transforms @compiled/react usages', () => {
     const actual = transform(`
       import { styled as styled2 } from '@compiled/react';
@@ -793,6 +797,101 @@ describe('styled tagged template expression', () => {
     expect(actual).toInclude('{color:red}');
   });
 
+  // This may seem unusual, but this occurs when @atlaskit/tokens (as of
+  // writing - v1.0.0) runs before @compiled/babel-plugin. @atlaskit/tokens'
+  // babel plugin will convert something like
+  //     ${token('color.background.accent.gray.subtler', colors.N20A)}
+  // to
+  //     ${`var(--ds-background-accent-gray-subtler, ${colors.N20A})`}
+  it('should transform variable within a nested template literal', () => {
+    const actual = transform(`
+      import { styled } from '@compiled/react';
+
+      const color = 'red';
+
+      const ListItem = styled.div\`
+        color: \${\`var(--my-variable, \${color})\`};
+      \`;
+    `);
+
+    expect(actual).toInclude('{color:var(--my-variable,red)}');
+  });
+
+  it('should transform variable within a nested template literal in hover selector', () => {
+    const actual = transform(`
+      import { styled } from '@compiled/react';
+
+      const color = 'red';
+
+      const ListItem = styled.div\`
+        &:hover {
+          background-color: \${\`var(--my-variable, \${color})\`};
+        }
+      \`;
+    `);
+
+    expect(actual).toInclude(':hover{background-color:var(--my-variable,red)}');
+  });
+
+  it('should transform variables within nested template literals in two properties', () => {
+    const actual = transform(`
+      import { styled } from '@compiled/react';
+
+      const color = 'red';
+      const color2 = 'blue';
+      const gridSize = 50;
+
+      const ListItem = styled.div\`
+        top: 5px;
+        right: \${gridSize}px;
+        color: \${\`var(--my-variable, \${color})\`};
+        border: 1px solid ${`var(--ds-border, \${color2})`};
+      \`;
+    `);
+
+    expect(actual).toIncludeMultiple([
+      '{top:5px}',
+      '{right:50px}',
+      '{color:var(--my-variable,red)}',
+      '{border:1px solid var(--ds-border,blue)}',
+    ]);
+  });
+
+  it('should transform variables within nested template literals within the same property', () => {
+    const actual = transform(`
+      import { styled } from '@compiled/react';
+
+      const color = 'red';
+      const color2 = 'blue';
+      
+      const ListItem = styled.div\`
+        background: linear-gradient(
+            \${\`var(--my-variable, \${color})\`},
+            \${\`var(--my-other-variable, \${color2})\`}
+        );
+      \`;
+    `);
+
+    expect(actual).toInclude(
+      '{background:linear-gradient(var(--my-variable,red),var(--my-other-variable,blue))}'
+    );
+  });
+
+  it('should transform variable within a heavily nested template literal', () => {
+    const actual = transform(`
+      import { styled } from '@compiled/react';
+
+      const color = 'red';
+      const interpolation = \`1px solid \${\`var(--my-variable, \${color})\`}\`;
+      
+      const ListItem = styled.div\`
+        border: \${interpolation};
+      \`;
+    `);
+
+    expect(actual).toInclude('{border:1px solid var(--my-variable,red)}');
+  });
+
   it('should handle destructuring in interpolation functions', () => {
     const actual = transform(`
       import { styled } from '@compiled/react';
@@ -817,6 +916,120 @@ describe('styled tagged template expression', () => {
       '__cmplp.loading ? colors.N50 : colors.N10',
       '__cmplp.loading ? colors.N100 : colors.N200',
     ]);
+  });
+
+  // For the next two test cases, the value of color and background-color
+  // cannot be statically evaluated, so the value gets extracted into
+  // the HTML style attribute
+
+  it('should transform variable within a template literal within an interpolation function', () => {
+    const actual = transform(
+      `
+      import { styled } from '@compiled/react';
+
+      const color = 'red';
+      const color2 = 'blue';
+
+      const ListItem = styled.div<{ ruleEnabled: boolean }>\`
+        color: \${({ ruleEnabled }) =>
+          (ruleEnabled ? color : \`var(--my-variable, \${color2})\`)};
+      \`;
+    `,
+      enableTypescript
+    );
+
+    expect(actual).toMatchInlineSnapshot(`
+      "const _2 = "._syazj09m{color:var(--my-variable,blue)}";
+      const _ = "._syaz5scu{color:red}";
+      const color = "red";
+      const color2 = "blue";
+      const ListItem = forwardRef(
+        ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+          const { ruleEnabled, ...__cmpldp } = __cmplp;
+          return (
+            <CC>
+              <CS>{[_, _2]}</CS>
+              <C
+                {...__cmpldp}
+                style={__cmpls}
+                ref={__cmplr}
+                className={ax([
+                  "",
+                  __cmplp.ruleEnabled ? "_syaz5scu" : "_syazj09m",
+                  __cmplp.className,
+                ])}
+              />
+            </CC>
+          );
+        }
+      );
+      "
+    `);
+  });
+
+  it('should transform variables within nested template literals that are all in an interpolation function', () => {
+    // this aims to represent the following use case:
+    //     background-color: ${({ isActive }) =>
+    //         `${
+    //             isActive
+    //                 ? `${token('some.token', color)}`
+    //                 : `${token('some.other.token', color2)}`
+    //         }`
+    //     };
+    //
+    // after token(...) has been processed by the babel plugin in
+    // @atlaskit/tokens (as of @atlaskit/tokens v1.0.0)
+    const actual = transform(
+      `
+      import { styled } from '@compiled/react';
+
+      const color = 'red';
+      const color2 = 'blue';
+
+      const ListItem = styled.div<{ isActive: boolean }>\`
+        background-color: \${({ isActive }) =>
+          \`\${
+            isActive
+              ? \`\${\`var(--my-variable, \${color})\`}\`
+              : \`\${\`var(--my-other-variable, \${color2})\`}\`
+          }\`
+        };
+      \`;
+    `,
+      enableTypescript
+    );
+
+    expect(actual).toMatchInlineSnapshot(`
+      "const _ = "._bfhkkqce{background-color:var(--_hudwc7)}";
+      const color = "red";
+      const color2 = "blue";
+      const ListItem = forwardRef(
+        ({ as: C = "div", style: __cmpls, ...__cmplp }, __cmplr) => {
+          const { isActive, ...__cmpldp } = __cmplp;
+          return (
+            <CC>
+              <CS>{[_]}</CS>
+              <C
+                {...__cmpldp}
+                style={{
+                  ...__cmpls,
+                  "--_hudwc7": ix(
+                    \`\${
+                      __cmplp.isActive
+                        ? \`\${\`var(--my-variable, \${color})\`}\`
+                        : \`\${\`var(--my-other-variable, \${color2})\`}\`
+                    }\`
+                  ),
+                }}
+                ref={__cmplr}
+                className={ax(["_bfhkkqce", __cmplp.className])}
+              />
+            </CC>
+          );
+        }
+      );
+      "
+    `);
   });
 
   it('should place classes in given order when static styles precede expression', () => {