+++ packages/babel-plugin/src/utils/build-styled-component.ts
@@ -16,8 +16,9 @@ import type { Metadata, Tag } from '../types';
 
 import { pickFunctionBody } from './ast';
 import { buildCssVariables } from './build-css-variables';
-import { compressClassNamesForAx } from './compress-class-names-for-ax';
+import { compressClassNamesForRuntime } from './compress-class-names-for-runtime';
 import { getItemCss } from './css-builders';
+import { getRuntimeClassNameLibrary } from './get-runtime-class-name-library';
 import { hoistSheet } from './hoist-sheet';
 import { applySelectors, transformCssItems } from './transform-css-items';
 import type { CSSOutput, CssItem } from './types';
@@ -169,7 +170,9 @@ const styledTemplate = (opts: StyledTemplateOpts, meta: Metadata): t.Node => {
           {...${hasInvalidDomProps ? DOM_PROPS_IDENTIFIER_NAME : PROPS_IDENTIFIER_NAME}}
           style={%%styleProp%%}
           ref={${REF_IDENTIFIER_NAME}}
-          className={ax([${classNames} ${PROPS_IDENTIFIER_NAME}.className])}
+          className={${getRuntimeClassNameLibrary(
+            meta
+          )}([${classNames} ${PROPS_IDENTIFIER_NAME}.className])}
         />
       </CC>
     );
@@ -240,7 +243,7 @@ export const buildStyledComponent = (tag: Tag, cssOutput: CSSOutput, meta: Metad
   const classNames = [
     ...[
       t.stringLiteral(
-        compressClassNamesForAx(
+        compressClassNamesForRuntime(
           uniqueUnconditionalCssOutput.classNames,
           meta.state.opts.classNameCompressionMap
         ).join(' ')