+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @compiled/utils
 
+## 0.8.0
+
+### Minor Changes
+
+- c4e6b7c0: Change TypeScript compiler target from es5 to es6.
+
 ## 0.7.0
 
 ### Minor Changes